[CHORES] Moved tests/IndonesiaMap into tests/app/components/IndonesiaMap
